+++ CoreAutomata/IScreenshotService.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace CoreAutomata
+{
+    public interface IScreenshotService : IDisposable
+    {
+        IPattern TakeScreenshot();
+    }
+}
