+++ FateGrandAutomata/Resources/values/preferences.xml
@@ -40,4 +40,5 @@
     <string name="pref_autoskill_cmd">autoskill_cmd</string>
     
     <string name="pref_ignore_notch">ignore_notch</string>
+    <string name="pref_use_root">use_root</string>
 </resources>
